moved examples to Extensions folder. scene cleanup, added trigger example
