+++ Assets/MixedRealityToolkit.Extensions/HandPhysicsService/Examples.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f415267b09628ac46b4af18997bea3c1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
