+++ Assets/MixedRealityToolkit.Extensions/HandPhysicsService/Examples/HandPhysicsServiceExample.unity.meta
