+++ Assets/MixedRealityToolkit.Extensions/HandPhysicsService/Examples/PhysicsTriggerEventReadout.cs.meta
