+++ Assets/MixedRealityToolkit.Extensions/HandPhysicsService/Interfaces.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 31f4c2d35c70fee4a9a036d8dcac4440
+guid: dea01b2d286630d4b821f96dd803c10e
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
