+++ Assets/MixedRealityToolkit.Extensions/HandPhysicsService/Microsoft.MixedReality.Toolkit.Extensions.HandPhysics.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f621f10f3b669bf40b1e4ff85dcfe2d8
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
