+++ Assets/MixedRealityToolkit.Extensions/HandPhysicsService/Profiles.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 72f3f20bcc74ac34c9e2d302529a8967
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
