+++ Assets/MixedRealityToolkit.Extensions/HandPhysicsService/Profiles/DefaultHandPhysicsServiceProfile.asset.meta
