+++ Assets/MixedRealityToolkit.Extensions/HandPhysicsService/Profiles/HandPhysicsMixedRealityRegisteredServiceProvidersProfile.asset
@@ -1,0 +1,24 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: eebbca41bb0b40d298ef201735d08616, type: 3}
+  m_Name: HandPhysicsMixedRealityRegisteredServiceProvidersProfile
+  m_EditorClassIdentifier: 
+  isCustomProfile: 1
+  configurations:
+  - componentType:
+      reference: Microsoft.MixedReality.Toolkit.Extensions.HandPhysics.HandPhysicsService,
+        Microsoft.MixedReality.Toolkit.Extensions.HandPhysics
+    componentName: HandPhysicsService
+    priority: 10
+    runtimePlatform: 8
+    configurationProfile: {fileID: 11400000, guid: 236baf35803f73342b69c842fe4afb1c,
+      type: 2}
