+++ Assets/MixedRealityToolkit.Extensions/HandPhysicsService/Profiles/HandPhysicsMixedRealityRegisteredServiceProvidersProfile.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 17db2de733b4d014e86d40cca58df36d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
