+++ Assets/MixedRealityToolkit.Extensions/HandPhysicsService/Profiles/HandPhysicsMixedRealityToolkitConfigurationProfile.asset.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 39e46c5684c168f4485eb54a2615f498
-folderAsset: yes
-DefaultImporter:
+guid: ec6c5962fb980c041854a25f84d6f2e6
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
